apps/media: change demand logic in consumer
This CL changes the way the consumer expresses demand to the upstream
producer. In effect, it establishes a minimum demand of one
outstanding packet. Previously, the demand expressed in the callback
to SupplyPacket was 0 outstanding packets because of the state of the
consumer at that moment. The demand would be updated in a few
milliseconds to 1 packet, requiring a separate round trip. This
effectively doubled the number of round trips required to deliver
a packet across a connection for most connections.

This CL also bumps up the video renderer's demand from 2 to 3
outstanding packets. This was done, because the renderer always keeps
around at least one packet, even if it's stale. This means that the
renderer's need for 2 fresh packets was not getting properly
expressed in starvation scenarios. In the future, the renderer
demand will be expressed in terms of PTS, so the packet retention
won't be an issue. For the moment, this additional demand incurs a
small cost in memory and latency.

+++ public/lib/media/transport/media_packet_consumer_base.h
@@ -110,7 +110,7 @@ class MediaPacketConsumerBase : public MediaPacketConsumer {
   // Called when a fatal error occurs. The default implementation does nothing.
   virtual void OnFailure();
 
-  size_t supplied_packets_outstanding() {
+  uint32_t supplied_packets_outstanding() {
     return counter_->packets_outstanding();
   }
 
@@ -167,7 +167,7 @@ class MediaPacketConsumerBase : public MediaPacketConsumer {
     }
 
     // Returns number of packets currently outstanding.
-    size_t packets_outstanding() {
+    uint32_t packets_outstanding() {
 #ifndef NDEBUG
       FTL_DCHECK(thread_checker_.IsCreationThreadCurrent());
 #endif
@@ -180,7 +180,7 @@ class MediaPacketConsumerBase : public MediaPacketConsumer {
     MediaPacketConsumerBase* owner_;
     // We keep the buffer set here, because it needs to outlive SuppliedPackets.
     SharedBufferSet buffer_set_;
-    size_t packets_outstanding_ = 0;
+    uint32_t packets_outstanding_ = 0;
 
 #ifndef NDEBUG
     ftl::ThreadChecker thread_checker_;
